requirements.txt
altgraph==0.16.1
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
fbs==0.8.4
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
gtasks==0.1.3
httplib2==0.13.1
idna==2.8
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter-client==5.3.1
jupyter-core==4.5.0
keyring==19.2.0
macholib==1.11
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.2
oauthlib==3.1.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
Pygments==2.4.2
PyInstaller==3.5
PyQt5==5.9.2
PyQt5-sip==4.19.19
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
tqdm==4.36.1
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
